+++ src/main/java/com/topay/common/config/session/SessionConfig.java
@@ -1,4 +1,4 @@
-package com.topay.common.config;
+package com.topay.common.config.session;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -10,7 +10,7 @@ import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactor
 import org.springframework.session.data.redis.config.annotation.web.http.EnableRedisHttpSession;
 
 @Configuration
-@EnableRedisHttpSession
+@EnableRedisHttpSession(maxInactiveIntervalInSeconds = 1800)
 public class SessionConfig {
 
     @Value("${spring.redis.session.host}")
